+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 * Added `ClientBuilder` for client initialization (#741).
 * [BREAKING] Updated client to Rust 2024 edition (#778).
 * [BREAKING] Removed `KeyStore` trait and added ability to provide signatures to `FilesystemKeyStore` and `WebKeyStore` (#744).
+* [BREAKING] Removed the `TransactionScriptBuilder` and associated errors from the `rust-client` (#781).
+* [BREAKING] Renamed "hash" with "commitment" for block headers, note scripts and accounts (#788, #789).
 
 ### Fixes